requirements.txt
huggingface-hub==0.16.4
langchain==0.0.284
langchainplus-sdk==0.0.20
langsmith==0.0.33
pandas==1.3.2
tokenizers==0.11.6
torch==1.12.1
torchvision==0.13.1
transformers==4.33.1
gradio==3.45.2
gradio_client==0.5.3
llama_cpp_python==0.2.7
pypdf==3.8.1
PyPDF2==3.0.1
python-dotenv==1.0.0
tiktoken==0.5.1
